Split up browser and server modules more clearly
