+++ packages/next/src/client/components/metadata/types.ts
@@ -1,0 +1,5 @@
+export type StreamingMetadataResolvedState = {
+  metadata: React.ReactNode
+  error: unknown | null
+  digest: string | undefined
+}
